backend/requirements.txt
fastapi
uvicorn
google-genai
requests
python-multipart
pydantic
python-dotenv
google.generativeai